ai-service-python/requirements.txt
fastapi
uvicorn
pymongo
redis
faiss-cpu
sentence-transformers
torch
numpy
pandas
pydantic
python-dotenv
requests
indic-nlp-library
nltk
transformers
coqui-tts
googletrans==4.0.0rc1
gtts
